+++ Classes/GHGist.h
@@ -10,6 +10,7 @@
 @property(nonatomic,strong)GHGistComments *comments;
 @property(nonatomic,strong)NSURL *htmlURL;
 @property(nonatomic,strong)NSDate *createdAtDate;
+@property(nonatomic,strong)NSDate *updatedAtDate;
 @property(nonatomic,strong)NSString *descriptionText;
 @property(nonatomic,strong)GHFiles *files;
 @property(nonatomic,readonly)GHUser *user;