+++ Classes/GistController.m
@@ -7,12 +7,14 @@
 #import "WebController.h"
 #import "GistController.h"
 #import "CodeController.h"
+#import "UserController.h"
 #import "CommentController.h"
 #import "CommentCell.h"
 #import "iOctocat.h"
 #import "NSDictionary+Extensions.h"
 #import "NSDate+Nibware.h"
 #import "SVProgressHUD.h"
+#import "LabeledCell.h"
 
 
 @interface GistController () <UIActionSheetDelegate>
@@ -20,15 +22,18 @@
 @property(nonatomic,weak,readonly)GHUser *currentUser;
 @property(nonatomic,readwrite)BOOL isStarring;
 @property(nonatomic,weak)IBOutlet UILabel *descriptionLabel;
-@property(nonatomic,weak)IBOutlet UILabel *numbersLabel;
-@property(nonatomic,weak)IBOutlet UILabel *ownerLabel;
+@property(nonatomic,weak)IBOutlet UILabel *forksCountLabel;
 @property(nonatomic,weak)IBOutlet UIImageView *iconView;
+@property(nonatomic,weak)IBOutlet UIImageView *forksIconView;
 @property(nonatomic,strong)IBOutlet UIView *tableHeaderView;
 @property(nonatomic,strong)IBOutlet UIView *tableFooterView;
 @property(nonatomic,strong)IBOutlet UITableViewCell *loadingCell;
 @property(nonatomic,strong)IBOutlet UITableViewCell *noFilesCell;
 @property(nonatomic,strong)IBOutlet UITableViewCell *loadingCommentsCell;
 @property(nonatomic,strong)IBOutlet UITableViewCell *noCommentsCell;
+@property(nonatomic,strong)IBOutlet LabeledCell *ownerCell;
+@property(nonatomic,strong)IBOutlet LabeledCell *createdCell;
+@property(nonatomic,strong)IBOutlet LabeledCell *updatedCell;
 @property(nonatomic,strong)IBOutlet CommentCell *commentCell;
 @end
 
@@ -74,7 +79,7 @@
 	if (!self.gist.comments.isLoaded) {
 		[self.gist.comments loadWithParams:nil success:^(GHResource *instance, id data) {
 			if (!self.gist.isLoaded) return;
-			NSIndexSet *sections = [NSIndexSet indexSetWithIndex:1];
+			NSIndexSet *sections = [NSIndexSet indexSetWithIndex:2];
 			[self.tableView reloadSections:sections withRowAnimation:UITableViewRowAnimationAutomatic];
 		} failure:^(GHResource *instance, NSError *error) {
 			[iOctocat reportLoadingError:@"Could not load the comments"];
@@ -107,10 +112,11 @@
 - (void)displayGist {
 	self.iconView.image = [UIImage imageNamed:(self.gist.isPrivate ? @"Private.png" : @"Public.png")];
 	self.descriptionLabel.text = self.gist.title;
-	if (self.gist.createdAtDate) {
-		self.ownerLabel.text = [NSString stringWithFormat:@"%@, %@", self.gist.user ? self.gist.user.login : @"unknown user", [self.gist.createdAtDate prettyDate]];
-		self.numbersLabel.text = self.gist.isLoaded ? [NSString stringWithFormat:@"%d %@", self.gist.forksCount, self.gist.forksCount == 1 ? @"fork" : @"forks"] : @"";
-	}
+	self.forksIconView.hidden = !self.gist.isLoaded;
+	self.ownerCell.contentText = self.gist.user.login;
+	self.createdCell.contentText = [self.gist.createdAtDate prettyDate];
+	self.updatedCell.contentText = [self.gist.updatedAtDate prettyDate];
+	self.forksCountLabel.text = self.gist.isLoaded ? [NSString stringWithFormat:@"%d %@", self.gist.forksCount, self.gist.forksCount == 1 ? @"fork" : @"forks"] : nil;
 }
 
 - (IBAction)addComment:(id)sender {
@@ -140,39 +146,57 @@
 #pragma mark TableView
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
-	return self.gist.isLoaded ? 2 : 1;
+	return self.gist.isLoaded ? 3 : 1;
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
 	if (!self.gist.isLoaded) return 1;
-	if (section == 0) return self.gist.files.isEmpty ? 1 : self.gist.files.count;
-	if (section == 1 && !self.gist.comments.isLoaded) return 1;
+	if (section == 0) return 3;
+	if (section == 1) return self.gist.files.isEmpty ? 1 : self.gist.files.count;
+	if (section == 2 && !self.gist.comments.isLoaded) return 1;
 	if (self.gist.comments.isEmpty) return 1;
 	return self.gist.comments.count;
 }
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section {
-	return (section == 1) ? @"Comments" : @"";
+	if (section == 1) {
+		return @"Files";
+	} else if (section == 2) {
+		return @"Comments";
+	} else {
+		return nil;
+	}
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
 	NSInteger section = indexPath.section;
 	NSInteger row = indexPath.row;
+	UITableViewCell *cell = nil;
 	if (self.gist.isLoading) return self.loadingCell;
-	if (section == 0 && self.gist.files.isEmpty) return self.noFilesCell;
-	static NSString *CellIdentifier = @"Cell";
-		UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:CellIdentifier];
-		if (cell == nil) {
-				cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
-		cell.textLabel.font = [UIFont systemFontOfSize:14.0];
-		}
 	if (section == 0) {
+		switch (row) {
+			case 0: cell = self.ownerCell; break;
+			case 1: cell = self.createdCell; break;
+			case 2: cell = self.updatedCell; break;
+			default: cell = nil;
+		}
+		BOOL isSelectable = row == 0 && [(LabeledCell *)cell hasContent];
+		cell.selectionStyle = isSelectable ? UITableViewCellSelectionStyleBlue : UITableViewCellSelectionStyleNone;
+		cell.accessoryType = isSelectable ? UITableViewCellAccessoryDisclosureIndicator : UITableViewCellAccessoryNone;
+	} else if (section == 1) {
+		if (self.gist.files.isEmpty) return self.noFilesCell;
+		static NSString *CellIdentifier = @"FileCell";
+		cell = [tableView dequeueReusableCellWithIdentifier:CellIdentifier];
+		if (cell == nil) {
+			cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
+			cell.textLabel.font = [UIFont systemFontOfSize:15.0];
+		}
 		NSDictionary *file = self.gist.files[row];
 		NSString *fileContent = [file safeStringForKey:@"content"];
 		cell.textLabel.text = [file safeStringForKey:@"filename"];
 		cell.selectionStyle = fileContent ? UITableViewCellSelectionStyleBlue : UITableViewCellSelectionStyleNone;
 		cell.accessoryType = fileContent ? UITableViewCellAccessoryDisclosureIndicator : UITableViewCellAccessoryNone;
-	} else if (section == 1) {
+	} else if (section == 2) {
 		if (!self.gist.comments.isLoaded) return self.loadingCommentsCell;
 		if (self.gist.comments.isEmpty) return self.noCommentsCell;
 		cell = [tableView dequeueReusableCellWithIdentifier:kCommentCellIdentifier];
@@ -187,11 +211,11 @@
 }
 
 - (UIView *)tableView:(UITableView *)tableView viewForFooterInSection:(NSInteger)section {
-	return section == 1 ? self.tableFooterView : nil;
+	return section == 2 ? self.tableFooterView : nil;
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
-	if (indexPath.section == 1 && self.gist.comments.isLoaded && !self.gist.comments.isEmpty) {
+	if (indexPath.section == 2 && self.gist.comments.isLoaded && !self.gist.comments.isEmpty) {
 		CommentCell *cell = (CommentCell *)[self tableView:tableView cellForRowAtIndexPath:indexPath];
 		return [cell heightForTableView:tableView];
 	}
@@ -199,7 +223,7 @@
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
-	if (section == 1) {
+	if (section == 2) {
 		return 56;
 	} else {
 		return 0;
@@ -208,8 +232,13 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
 	if (!self.gist.isLoaded) return;
-	if (indexPath.section == 0) {
-		CodeController *codeController = [[CodeController alloc] initWithFiles:self.gist.files currentIndex:indexPath.row];
+	NSInteger section = indexPath.section;
+	NSInteger row = indexPath.row;
+	if (section == 0 && row == 0 && self.gist.user) {
+		UserController *userController = [[UserController alloc] initWithUser:self.gist.user];
+		[self.navigationController pushViewController:userController animated:YES];
+	} if (section == 1) {
+		CodeController *codeController = [[CodeController alloc] initWithFiles:self.gist.files currentIndex:row];
 		[self.navigationController pushViewController:codeController animated:YES];
 	}
 }